sql_bolt/requirements.txt
atlassian-python-api
python-dotenv
langchain
langchain-openai
slack_bolt
langchain-pinecone
pandas
google-cloud-bigquery
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pyarrow
db-dtypes